Move samples to isolated package to verify accessibility issues
