+++ src/test/scala/info/fingo/sample/spata/SampleITS.scala
@@ -3,16 +3,17 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package info.fingo.spata.sample
+package info.fingo.sample.spata
 
-import java.nio.file.{Files, Paths}
-import scala.io.Codec
 import cats.effect.IO
 import cats.effect.unsafe.implicits.global
 import fs2.Stream
-import org.scalatest.funsuite.AnyFunSuite
-import info.fingo.spata.{CSVParser, CSVRenderer}
 import info.fingo.spata.io.{Reader, Writer}
+import info.fingo.spata.{CSVParser, CSVRenderer}
+import org.scalatest.funsuite.AnyFunSuite
+
+import java.nio.file.{Files, Paths}
+import scala.io.Codec
 
 // Sample from readme (fahrenheit to celsius conversion), in form of test
 class SampleITS extends AnyFunSuite {
